requirements.txt
asgiref==3.3.1
astroid==2.5.2
autopep8==1.5.6
Django==3.1.7
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-taggit==1.3.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.7.0
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
